2021-01-16 - SupportShouldProcess
Added SupportsShouldProcess to the following Cmdlets:
* Publish-DSClientEncryptionKeys
* Register-DSClient
* Set-DSClientAdvancedConfig
* Set-DSClientConfiguration
* Set-DSClientDefaultConfiguration
* Set-DSClientMSSqlServerBackupSet (NOTE: Common Backup Set Parameters currently do NOT support "ShouldProcess")
* Set-DSClientRegistration
* Set-DSClientRetentionRule
* Set-DSClientSMTPNotification
* Set-DSClientSNMPHeartbeat
* Set-DSClientUser
* Set-DSClientUserGroup
* Stop-DSClientActivity
* Unpublish-DSClientEncryptionKeys

Updated ShouldProcess on the following Cmdlets:

* Set-DSClientUserGroupRole

+++ PSAsigraDSClient/UnpublishDSClientEncryptionKeys.cs
@@ -3,7 +3,7 @@ using AsigraDSClientApi;
 
 namespace PSAsigraDSClient
 {
-    [Cmdlet(VerbsData.Unpublish, "DSClientEncryptionKeys")]
+    [Cmdlet(VerbsData.Unpublish, "DSClientEncryptionKeys", SupportsShouldProcess = true)]
 
     public class UnpublishDSClientEncryptionKeys: DSClientCmdlet
     {
@@ -11,11 +11,14 @@ namespace PSAsigraDSClient
         {
             ClientConfiguration DSClientConfigMgr = DSClientSession.getConfigurationManager();
 
-            WriteVerbose("Performing Action: Set Encryption Key Forwarding to False");
-            DSClientConfigMgr.setForwardingEncryptionKey(false);
+            if (ShouldProcess("DS-System", "Remove DS-Client Encryption Keys"))
+            {
+                WriteVerbose("Performing Action: Set Encryption Key Forwarding to False");
+                DSClientConfigMgr.setForwardingEncryptionKey(false);
 
-            WriteVerbose("Performing Action: Remove Encryption Keys from DS-System");
-            DSClientConfigMgr.deleteEncryptionKeyFromDSSystem();
+                WriteVerbose("Performing Action: Remove Encryption Keys from DS-System");
+                DSClientConfigMgr.deleteEncryptionKeyFromDSSystem();
+            }
 
             DSClientConfigMgr.Dispose();
         }
